+++ adaptive-backend/internal/services/providers/deepseek/stream_adapter.go
@@ -1,4 +1,4 @@
-package providers
+package deepseek
 
 import (
 	"context"
