+++ adaptive-backend/internal/services/providers/llm_provider.go
@@ -2,6 +2,7 @@ package providers
 
 import (
 	"adaptive-backend/internal/services/providers/anthropic"
+	"adaptive-backend/internal/services/providers/deepseek"
 	"adaptive-backend/internal/services/providers/openai"
 	"adaptive-backend/internal/services/providers/provider_interfaces"
 	"errors"
@@ -24,7 +25,7 @@ func NewLLMProvider(providerName string) (provider_interfaces.LLMProvider, error
 		}
 		return service, nil
 	case "deepseek":
-		service, err := NewDeepSeekService()
+		service, err := deepseek.NewDeepSeekService()
 		if err != nil {
 			return nil, err
 		}
